pyproject.toml
[project]
name = "skill-atlas"
version = "0.0.1"
description = "Procedural skill-atlas generator"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "networkx>=3.0",
    "attrs>=23.1.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "ruff",
    "pyright",
    "pytest",
    "pytest-cov",
    "attrs>=23.1.0",
    "PyYAML>=6.0",
]

[project.scripts]
skill-atlas = "skill_atlas.cli.main:run"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.pyright]
typeCheckingMode = "strict"
extraPaths = [".venv/lib/python3.12/site-packages"]
